# Refs #28459 -- Improved performance of BaseExpression.convert_value().