@timgraham 2017-08-08
`Expression.convert_value()` is documented so I don't think this change is acceptable:
`ref/models/expressions.txt:    .. method:: convert_value(value, expression, connection)`