@sir-sigurd 2017-08-08
I think I will propose to deprecate `convert_value()` in the future but for now I'd like to not break it.
If you think that the current approach is breaking I could rework it in compatible way.