@timgraham 2017-08-08
The current approach changes `convert_value()` from a method accepting arguments to a cached property, so Expression subclasses overriding it will break, correct?