@sir-sigurd 2017-08-08
But this property returns callable. For example:
```
In [6]: class A:
   ...:     @cached_property
   ...:     def method(self):
   ...:         return lambda arg: 'property ' + arg
   ...:
In [11]: class B(A):
    ...:     def method(self, arg):
    ...:         return 'method ' + arg + ' ' + super().method(arg)
    ...:
In [12]: B().method('test')
Out[12]: 'method test property test'
```

I see the only difference in behavior: currently `Expression.convert_value()` is a function that could be called if supply instance to along with other arguments.