@timgraham 2017-08-08
I see now (and of course there are expressions in Django that show backwards compatibility is okay). Well, perhaps it would be good to add some comments explaining why things are implemented the way they are because IMO this obscures the implementation quite a bit. If we proceed with the patch, it might be useful to avoid the repetition of `lambda value, expression, connection: None if value is None` if possible. \cc @jarshwah for input.